Issue 611: fix warnings.  Also added in all the .h files to the CMake files, so that IDEs like VS can find them. 

Most of the fixes involved int/uint/size_t conversions, but there were a few actual bugs hiding behind all the other warnings.

+++ source/Dictionary.cpp
@@ -30,7 +30,7 @@ bool BasicDictionary::hasKey(const std::string& key) const
     return items.find(key) != items.end();
 }
 
-int  BasicDictionary::deleteItem(const std::string& key)
+size_t  BasicDictionary::deleteItem(const std::string& key)
 {
     return items.erase(key);
 }
