+++ source/rrRoadRunnerOptions.h
@@ -195,7 +195,7 @@ namespace rr
 		/**
 		* remove a value
 		*/
-		virtual int deleteItem(const std::string& key);
+		virtual size_t deleteItem(const std::string& key);
 
 		/**
 		* list of keys in this object.
